+++ src/main/java/com/novi/poffinhouse/repositories/PokemonRepository.java
@@ -3,5 +3,5 @@ package com.novi.poffinhouse.repositories;
 import com.novi.poffinhouse.models.pokemon.Pokemon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokemonRepository extends JpaRepository <Pokemon,int> {
+public interface PokemonRepository extends JpaRepository <Pokemon,Long> {
 }
