+++ src/main/java/com/novi/poffinhouse/repositories/TeamRepository.java
@@ -3,5 +3,5 @@ package com.novi.poffinhouse.repositories;
 import com.novi.poffinhouse.models.pokemon.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeamRepository extends JpaRepository<Team, int> {
+public interface TeamRepository extends JpaRepository<Team, Long> {
 }
